client :: moved notes component in the domains folder
